commit 002

+++ tutorial-projects/basiccalculator.c
@@ -4,13 +4,13 @@
 int main()
 {
     // Let's create a simple calculator that can sum two integers
-    int num1;
-    int num2;
+    double num1;
+    double num2;
     printf("Enter first number: ");
-    scanf("%d", &num1);
+    scanf("%lf", &num1);
     printf("Enter second number: ");
-    scanf("%d", &num2);
-    printf("Answer: %d", num1 + num2);
+    scanf("%lf", &num2);
+    printf("Answer: %lf", num1 + num2);
 
     return 0;
 }
